Refactor save passphrase modal into a component #189
