+++ src/components/passphrase/savePassphrase.less
@@ -1,4 +1,4 @@
-.dialog-save {
+save-passphrase {
   .missing {
     padding: 0 5px 0;
     font-weight: bold;
