+++ src/components/passphrase/savePassphrase.pug
@@ -1,24 +1,23 @@
-md-dialog.dialog-save(flex='80', aria-label='Save your passphrase in a safe place!')
-  form
-    md-toolbar
-      .md-toolbar-tools
-        h2
-          span(ng-if="!$ctrl.enter") Save your passphrase in a safe place!
-          span(ng-if="$ctrl.enter") Enter the missing word to continue
-    md-dialog-content
-      .md-dialog-content
-        md-input-container.md-block(ng-if="!$ctrl.enter")
-          textarea.passphrase(ng-bind='$ctrl.passphrase', md-autofocus, readonly)
-        div(ng-if="$ctrl.enter")
-          p.passphrase
-            span {{ $ctrl.pre }}
-            span.missing -----
-            span {{ $ctrl.pos }}
-          md-input-container.md-block(md-is-error='!$ctrl.missing_ok')
-            label Enter the missing word
-            input(ng-model='$ctrl.missing_input')
-    md-dialog-actions(layout='row')
-      md-button(ng-click="$ctrl.close()") Close
-      span(flex)
-      md-button.yes-its-save-button(ng-click="$ctrl.next()", ng-show='!$ctrl.enter') Yes! It's safe!
-      md-button.ok-button(ng-click="$ctrl.ok()", ng-show='$ctrl.enter', ng-disabled='!$ctrl.missing_ok') OK
+form
+  md-toolbar
+    .md-toolbar-tools
+      h2
+        span(ng-if="!$ctrl.enter") Save your passphrase in a safe place!
+        span(ng-if="$ctrl.enter") Enter the missing word to continue
+  md-dialog-content
+    .md-dialog-content
+      md-input-container.md-block(ng-if="!$ctrl.enter")
+        textarea.passphrase(ng-bind='$ctrl.passphrase', md-autofocus, readonly)
+      div(ng-if="$ctrl.enter")
+        p.passphrase
+          span {{ $ctrl.pre }}
+          span.missing -----
+          span {{ $ctrl.pos }}
+        md-input-container.md-block(md-is-error='!$ctrl.missing_ok')
+          label Enter the missing word
+          input(ng-model='$ctrl.missing_input')
+  md-dialog-actions(layout='row')
+    md-button(ng-click="$ctrl.close()") Close
+    span(flex)
+    md-button.yes-its-save-button(ng-click="$ctrl.next()", ng-show='!$ctrl.enter') Yes! It's safe!
+    md-button.ok-button(ng-click="$ctrl.ok()", ng-show='$ctrl.enter', ng-disabled='!$ctrl.missing_ok') OK
